+++ src/lib.rs
@@ -38,13 +38,13 @@
 //! Create application with anchoring service
 //!
 //! ```rust,no_run
-//! use exonum_cli::NodeBuilder;
+//! use exonum_cli::{NodeBuilder, Spec};
 //!
 //! #[tokio::main]
-//! async fn main() -> Result<(), failure::Error> {
+//! async fn main() -> Result<(), anyhow::Error> {
 //!     exonum::helpers::init_logger()?;
 //!     NodeBuilder::new()
-//!         .with_rust_service(exonum_btc_anchoring::BtcAnchoringService)
+//!         .with(Spec::new(exonum_btc_anchoring::BtcAnchoringService))
 //!         .run()
 //!         .await
 //! }